Cross-validated ANOVA for O-PLS models

Significance testing for OPLS models

cvanova(smod)

Arguments

smod
OPLS_metabom8 object of the package metabom8.

Value

data.frame describing ANOVA stats incl, p value

Details

The CV-ANOVA diagnostic formally compares the fit of two models to the same data by the size of their residuals. The function tests the residuals of the linear regression between cross-validated scores of the predictive O-PLS component and the response Y, with the variation of Y around its mean. The p value is derived from an F-test with the null hypothesis of equal residuals of the two models. For detailed information on p value colculation see refrence further below.

References

Eriksson, L, et al. (2008) CV-ANOVA for significance testing of PLS and OPLS models. Journal of Chemometrics, 22, 594-600.

See also

Author

torben.kimhofer@murdoch.edu.au