Cliff's delta (effect size)

Non-parametric effect size estimate used to quantify the degree to which two variable distributions overlap. Output is scaled to range between -1 to 1, with the arithmetic sign indicating if the comparator group lies to the right (higher) or left (lower) to a the reference.

es_cdelta(ref, comp)

Arguments

ref
num, reference group

comp
num, comparator group

Value

This function returns Cliff's delta effect size ranging from -1 to 1

References

Cliff, N (1993), Dominance statistics: Ordinal analyses to answer ordinal quetsions Psychological Bulletin, 114.(3), 494-509.https://doi.org/10.1037%2F0033-2909.114.3.494

See also

Author

torben.kimhofer@murdoch.edu.au

torben.kimhofer@murdoch.edu.au

Examples

# define two distrubutions
a=rnorm(100, mean=0, sd=1)
b=rnorm(100, mean=2, sd=1)

# Cliff's delta
 es_cdelta(ref=a, comp=b)

#> [1] 0.8494

 es_cdelta(ref=b, comp=a)

#> [1] -0.8494