Excise standard chemical shift regions

Excise standard chemical shift regions

excise1d(X, ppm)

Arguments

X
num matrix: NMR data, spectra in rows

ppm
ppm, num array - chemical shift positions, length matches to columns in X

Value

The function exports the following two objects into the currently active R environment (no variable assignments needed):

Xc, num matrix: column reduced

ppc, num array - chemical shift positions, length is equal to ncol(X_c)

Details

Function removes the following chemical shift reagions from 1H NMR spectra acquired for small molecule research:

min(ppm) to 0.25 ppm (cap upfield)

4.5 - 5.2 ppm (residual water)

5.5 - 6.0 ppm (urea)

9.7 to max(ppm) (cap downfield)

NA

See also

Author

torben.kimhofer@murdoch.edu.au