Plotting PCA or OPLS loadings

Plotting PCA or OPLS loadings

plotload(
  mod,
  shift = c(0, 10),
  pc = 1,
  type = "Backscaled",
  title = NULL,
  r_scale = F
)

Arguments

mod
PCA or OPLS model generated via metabom8 package functions.

shift
ppm region to visualise.

pc
index of principal component to visualise, set to 1 if input model is OPLS

type
Type of loadings visualisation, either 'Statistical reconstruction' or 'Backscaled' (see Details).

title
Plot title.

r_scale
logical, adjust limits of color gradient to 0 and 1 (only applies for type stat reconstruction)

Details

OPLS: If type='Statistical reconstruction' the function calculates the covariance (y axis) and Pearson's correlation (colouring) of the predictive OPLS scores with each X variable (x axis is ppm variable). If type='Backscaled' the OPLS loadings are backscaled with X feature standard deviations. Results are plotted over ppm, coloured according to OPLS model weights. Often, the latter method visualises model importance more robust due to the presence of false positive correlations. PCA: Function always calculates the statistical recostruction.

NA

References

Cloarec, O., et al. (2005). Evaluation of the Orthogonal Projection on Latent Structure Model Limitations Caused by Chemical Shift Variability and Improved Visualization of Biomarker Changes in 1H NMR Spectroscopic Metabonomic Studies. Analytical Chemistry 77.2, 517-26.

See also