Spectral calibration to a chemical shift reference (1D NMR)

Spectral calibration to a chemical shift reference.

calibrate(X, ppm, type = "glucose")

Arguments

X
num matrix, NMR data matrix with rows representing spectra and columns representing chemical shift position

ppm
num vector, matched to columns of X

type
str or num. Str: Either 'tsp' or 'glucose' for urine or blood-derived spectra, respectively (see Details). Num: ppm range of max height signal that will be used to reference to zero

Value

num matrix X, calibrated NMR data matrix.

Details

Spectral calibration to a chemical shift reference. If type='tsp' calibration will be performed using highest peak in interval -0.2 ppm to +0.2 ppm 0 ppm (Trimethylsilylpropanoic acid resonance). If type='glucose', the glucose doublet (alphanomeric proton with J const within 3.6 and 4.2 Hz) will be centered at 5.23 pmm , originating from the alpha anomer of glucose, will be used for calibration. Blood serum-derived spectra can also be calibrated with input argument type='Plasma'.

References

Dona, A.C., et al. (2014) Precision high-throughput proton NMR spectroscopy of human urine, serum, and plasma for large-scale metabolic phenotyping. Analytical Chemistry. 86.19. 9887-94.

Author

torben.kimhofer@gmail.com

Examples

data(covid_raw)
matspec(X, ppm, shift=c(-0.1,0.1))

#> Warning: minimal value for n is 3, returning requested palette with 3 different levels

#> Warning: minimal value for n is 3, returning requested palette with 3 different levels

X_tsp=calibrate(X, ppm, type='tsp')
matspec(X, ppm, shift=c(-0.1,0.1))

#> Warning: minimal value for n is 3, returning requested palette with 3 different levels

#> Warning: minimal value for n is 3, returning requested palette with 3 different levels

# calibrate with glucose (blood plasma)
matspec(X, ppm, shift=c(5.15, 5.3))

#> Warning: minimal value for n is 3, returning requested palette with 3 different levels

#> Warning: minimal value for n is 3, returning requested palette with 3 different levels

X_tsp=calibrate(X, ppm, type='glucose')
matspec(X, ppm, shift=c(5.15, 5.3))

#> Warning: minimal value for n is 3, returning requested palette with 3 different levels

#> Warning: minimal value for n is 3, returning requested palette with 3 different levels