Full width at half-maximum (fwhm)

Calculating full width at half maximum (FWHM, aka line width). This function returns the ppm difference where peak line crosses half of the peak height. It requires one signal across all spectra within ppm ranges specified in shift.

lw(X, ppm, shift = c(-0.1, 0.1), sf)

Arguments

X
num matrix, NMR data with rows representing spectra.

ppm
num array describing chemical shift positions, its length equals to nrow(X).

shift
num array(2), chemical shift range of a singlet for which fwhm is calculated

sf
num, operating frequency of the spectrometer (meta$SFO1)

Value

Array of line widths in ppm. To convert from ppm to Hertz (Hz), multiply values with the spectrometer frequency (column a_SF01 in meta data frame).

NA

See also

Author

torben.kimhofer@murdoch.edu.au

Examples

load(covid)

#> Error in load(covid): object 'covid' not found

lw(X, ppm) # in ppm

#> Error in lw(X, ppm): argument "sf" is missing, with no default

lw(X, ppm) * meta$a_SFO1 # in Hz

#> Error in lw(X, ppm): argument "sf" is missing, with no default