Read-in FIDs and process to spectra

Read-in FIDs and process to spectra

read1d_raw(
  path,
  exp_type = list(exp = c("PROF_PLASMA_CPMG128_3mm"), pulprog = c("noesygppr1d")),
  apodisation = list(fun = "exponential", lb = 0.2),
  zerofil = 1L,
  return = "absorption",
  verbose = 1,
  recursive = TRUE,
  n_max = 1000,
  filter = TRUE
)

Arguments

path
char, path to file directory containing spectra

exp_type
named list, filter for acquisition paramters of experiments to read-in (see Details)

apodisation
named list, apodisation function and its parameters (see Details)

zerofil
int, amount of zeros to append to FID given as exponent added of base 2 (see Details)

return
char, return mode of spectra: absorption, dispersion or magnitude mode

verbose
num, different verbose levels: 0 (no info), 1 (overview), 2 (detailed), 3 (step-by-step for debugging)

recursive
logic, if TRUE recursively search all subfolders of path for specified NMR files

n_max
int, maximum number of experiments to read-in

filter
logic, remove experiments with incomplete file systems (TRUE is recommended)

Value

The function exports the following three objects into the currently active R environment (no variable assignments needed):

X, num matrix: NMR data, spectra in rows

ppm, num array - chemical shift positions, length matches to columns in X

meta, data.frame - spectrometer metadata as extracted from individual acqus files, row-matched to X

Objects in the R environment with the same variable names will be overwritten.

Details

In the first step, read-in are FIDs generated with experimental condition(s) specified with the exp_type argument. This represents a list with each element representing a parameter condition, named according to spectrometer parameters listed in acqus file. For example, to read standard 1D NMR experiments use exp_type=list(exp='noesygppr1d'). More than one argument can be provided as list element.

The apodisation argument is a named list specifying the function name in element fun and functions-specific paramter arguments. There are the following different apodisation functions and arguments: #' @return

exponential, arguments: lb (line broadening factor)

cosine, no further arguments

sine, no further arguments

sem, combined sine-bell - exponential fct: arguments: lb (line broadening factor)

The zerofil argument specifies the amount of zeros to append to the FID and is expressed as exponent addand in the binary numeral system: 2^(1+x), with x being the zerofil parameter argument. Hence, zerofil=1 doubles the amount of data points.

NA

See also

Author

torben.kimhofer@murdoch.edu.au

Examples

path<-system.file('extdata/',  package = 'metabom8')
read1d_raw(path,  exp_type=list(exp='PROF_PLASMA_NOESY'), apodisation=list(fun='exponential', lb=0.2), n_max=3)

#> Error in .filterExp_files(pars, exp_type, f_list): argument "n_max" is missing, with no default